Implementation the level of access and give authority to users

+++ Modules/Mlk/Category/Http/Controllers/CategoryController.php
@@ -2,16 +2,16 @@
 
 namespace Mlk\Category\Http\Controllers;
 
-use Mlk\Category\Http\Requests\CategoryRequest;
 use Mlk\Category\Models\Category;
-use Mlk\Category\Repositories\CategoryRepo;
 use Mlk\Category\Services\CategoryService;
-use Illuminate\Routing\Controller as BaseController;
+use Mlk\Category\Repositories\CategoryRepo;
+use Mlk\Category\Http\Requests\CategoryRequest;
+use App\Http\Controllers\Controller;
 
-class CategoryController extends BaseController
+class CategoryController extends Controller
 {
     public CategoryRepo $repo;
-    
+
     public CategoryService $service;
 
     public function __construct(CategoryRepo $categoryRepo, CategoryService $categoryService)
@@ -23,20 +23,26 @@ class CategoryController extends BaseController
 
     public function index()
     {
-        $categories = $this->repo->index()->paginate(10); 
+        $this->authorize('index', Category::class);
+
+        $categories = $this->repo->index()->paginate(10);
 
         return view('Category::index', compact('categories'));
     }
 
     public function create()
     {
-        $categories = $this->repo->index()->get(); 
+        $this->authorize('index', Category::class);
+
+        $categories = $this->repo->index()->get();
 
         return view('Category::create', compact('categories'));
     }
 
     public function store(CategoryRequest $request)
     {
+        $this->authorize('index', Category::class);
+
         $this->service->store($request);
 
         return to_route('categories.index')->with(['success_message' => 'دسته بندی با موفقیت ساخته شد']);
@@ -44,29 +50,37 @@ class CategoryController extends BaseController
 
     public function edit($id)
     {
-        $category = $this->repo->findById($id); 
+        $this->authorize('index', Category::class);
 
-        $categories = $this->repo->index()->where('id', '!=', $category->id)->get();  
+        $category = $this->repo->findById($id);
+
+        $categories = $this->repo->index()->where('id', '!=', $category->id)->get();
 
         return view('Category::edit', compact(['category', 'categories']));
     }
 
     public function update(CategoryRequest $request, $id)
     {
-        $this->service->update($request, $id); 
+        $this->authorize('index', Category::class);
+
+        $this->service->update($request, $id);
 
         return to_route('categories.index')->with(['success_message' => 'دسته بندی با موفقیت ابدیت شد']);
     }
 
     public function destroy($id)
     {
-        $this->repo->delete($id); 
+        $this->authorize('index', Category::class);
+
+        $this->repo->delete($id);
 
         return to_route('categories.index')->with(['success_message' => 'دسته بندی با موفقیت حذف شد']);
     }
 
     public function changeStatus($id)
     {
+        $this->authorize('index', Category::class);
+
         $category = $this->repo->findById($id);
         $this->repo->changeStatus($category);
