+++ Modules/Mlk/User/Http/Controllers/UserController.php
@@ -3,14 +3,16 @@
 namespace Mlk\User\Http\Controllers;
 
 
-use Mlk\User\Http\Requests\UserRequest;
-use Mlk\User\Http\Requests\UserUpdateRequest;
 use Mlk\User\Models\User;
-use Mlk\User\Repositories\UserRepo;
 use Mlk\User\Services\UserService;
-use Illuminate\Routing\Controller as BaseController;
+use Mlk\Role\Repositories\RoleRepo;
+use Mlk\User\Repositories\UserRepo;
+use Mlk\User\Http\Requests\UserRequest;
+use Mlk\User\Http\Requests\AddRoleRequest;
+use Mlk\User\Http\Requests\UserUpdateRequest;
+use App\Http\Controllers\Controller;
 
-class UserController extends BaseController
+class UserController extends Controller
 {
 
     public UserRepo $repo;
@@ -24,6 +26,8 @@ class UserController extends BaseController
 
     public function index()
     {
+        $this->authorize('index',User::class);
+
         $users = $this->repo->index();
 
         return view('User::index', compact('users'));
@@ -31,11 +35,15 @@ class UserController extends BaseController
 
     public function create()
     {
+        $this->authorize('index',User::class);
+
         return view('User::create');
     }
 
     public function store(UserRequest $request)
     {
+        $this->authorize('index',User::class);
+
         $this->service->store($request);
 
         return to_route('users.index');
@@ -43,6 +51,8 @@ class UserController extends BaseController
 
     public function edit($id)
     {
+        $this->authorize('index',User::class);
+
         $user = $this->repo->findById($id);
 
         return view('User::edit', compact('user'));
@@ -50,6 +60,8 @@ class UserController extends BaseController
 
     public function update(UserUpdateRequest $request, $id)
     {
+        $this->authorize('index',User::class);
+
         $this->service->update($request, $id);
 
         return to_route('users.index');
@@ -57,9 +69,43 @@ class UserController extends BaseController
 
     public function destroy($id)
     {
+        $this->authorize('index',User::class);
+
         $this->repo->delete($id);
 
         return to_route('users.index')->with(['success_delete' => 'کاربر با موفقیت حذف شد']);
+    }
+
+    public function addRole($user_id, RoleRepo $roleRepo)
+    {
+        $this->authorize('index',User::class);
+
+        $roles = $roleRepo->index()->get();
+
+        return view('User::add-roles', compact(['user_id', 'roles']));
+    }
+
+    public function addRoleStore(AddRoleRequest $request, $userId)
+    {
+        $this->authorize('index',User::class);
+
+        $user = $this->repo->findById($userId);
+        $this->service->addRole($request->role, $user);
+
+        alert()->success('اضافه کردن مقام به کاربر', 'عملیات با موفقیت انجام شد');
+        return to_route('users.index');
+    }
+
+    public function removeRole($userId, $roleId, RoleRepo $roleRepo)
+    {
+        $this->authorize('index',User::class);
+
+        $user = $this->repo->findById($userId);
+        $role = $roleRepo->findById($roleId);
+        $this->service->deleteRole($user, $role->name);
+
+        alert()->success('حذف کردن مقام', 'عملیات با موفقیت انجام شد');
+        return to_route('users.index');
     }
 
 }
